# Conflicts:
#	src/main/java/com/example/parkinglot/config/SecurityConfiguration.java
#	src/main/webapp/app/login/login.component.html

+++ src/main/java/com/example/parkinglot/service/PriceService.java
@@ -6,6 +6,7 @@ import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.util.Optional;
 
@@ -16,7 +17,7 @@ public class PriceService {
     private PriceRepository priceRepository;
 
     @Transactional
-    public Price updatePriceByDurationAndValue(int duration, BigInteger newPrice) {
+    public Price updatePriceByDurationAndValue(int duration, BigDecimal newPrice) {
         Optional<Price> optPrice = priceRepository.findByDuration(duration);
         if (optPrice.isPresent()) {
             Price price = optPrice.get();
